Implemented interruptless processing and some bugfixes
- Fixed bug that scetch could not be compiled (using VisualMicro)
because Wire.h was not found. Arduino-Libraries must be also included at
top of the *.ino file to prevent this problem.
- Fixed bug that first interrupt was ignored
- Added option LOWLEVELCONFIG_GYRO_USE_INTERRUPT. If set to false, the
interrupt signal is ignored. Data is processed if the FIFO buffer has
enough bytes and the interrupt status is set to "data ready". If set to
true, also the interrupt signal itself must be received to enter the
processing method.
- Minor modifications in Quaternion class and config.h, debugging is now
enabled/disabled using the flag LOWLEVELCONFIG_ENABLE_DEBUGGING (set to
true or false).