Improved SBUS parser, implemented basic ESC output
- The parser for the SBUS protocoll has been improved, it is now able to
resync on the start byte even if there is another byte in the data
stream which has the same value, but actually is a data byte

- Implemented a basic ESC throttle output calculation. For now, it has
no gyro feedback loop or PID governor, it simply uses the
throttle/pitch/roll/yaw input (RC / sbus input).
This is just for testing, however, if all 4 motors and ESCs are
identically calibrated and we tune some parameters, it should be
possible to control the quadrocopter like a  non-stabilized helicopter.