Added "Armed/Disarmed" mode+LED and AutoRotate-Overwrite
Added armed/disarmed mode. In Disarmed mode (new LED at pin 12 -> dark),
all motor throttles are set to 0 and no inputs are processed (except
detection to arm).
In armed mode, all inputs are processed and all motor throttles are set
at least to 0.03. If throttle signal from receiver is 0.0 for 5 seconds,
system is disarmed. If autorotation switch is activated, system is
disarmed immediately. To arm a disarmed system, move YAW-stick to the
left and ROLL-stick to the right while throttle is at 0.0.